custom authentication done
finished custom authentication. the main function in fuzzer.py now runs
discover(args). the authentication is able to get request for dvwa and
bwapp websites. the other links in that page does not seem to matter.
the name of the folder "Requests" should be renamed "requests" since
it's what the init actually calls, but I think I will screw up the repo
if I do something like that, so let's do it the next time we meet.